onboarding/requirements.txt
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==39.0.1
fire==0.5.0
Flask==2.3.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
prometheus-client==0.17.0
pycparser==2.21
pycryptodomex==3.17
PyNaCl==1.5.0
requests==2.31.0
six==1.16.0
termcolor==2.3.0
urllib3==2.0.3
Werkzeug==2.3.6
